requirements.txt
black==24.3.0
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
flasgger==0.9.7.1
numpy==1.26.4
pandas==2.2.1
psycopg2-binary==2.9.9
pylint==3.1.0
python-dotenv==1.0.1
scipy==1.12.0
Sphinx==7.2.6
sphinx-rtd-theme==2.0.0
SQLAlchemy==2.0.29